+++ components/keyboard-data.js
@@ -1,0 +1,81 @@
+const keysObj = {
+  Backquote: ['§', '±', '>', '<'],
+  Digit1: ['1', '!', '1', '!'],
+  Digit2: ['2', '@', '2', '"'],
+  Digit3: ['3', '#', '3', '№'],
+  Digit4: ['4', '$', '4', '%'],
+  Digit5: ['5', '%', '5', ':'],
+  Digit6: ['6', '^', '6', ','],
+  Digit7: ['7', '&', '7', '.'],
+  Digit8: ['8', '*', '8', ';'],
+  Digit9: ['9', '(', '9', '('],
+  Digit0: ['0', '+', '0', ')'],
+  Minus: ['-', '_', '-', '_'],
+  Equal: ['=', '+', '=', '+'],
+  Backspace: ['Backspace', 'Backspace', 'Backspace', 'Backspace'],
+  Tab: ['Tab', 'Tab', 'Tab', 'Tab'],
+  KeyQ: ['q', 'q', 'й', 'й'],
+  KeyW: ['w', 'w', 'ц', 'ц'],
+  KeyE: ['e', 'e', 'у', 'у'],
+  KeyR: ['r', 'r', 'к', 'к'],
+  KeyT: ['t', 't', 'е', 'е'],
+  KeyY: ['y', 'y', 'н', 'н'],
+  KeyU: ['u', 'u', 'г', 'г'],
+  KeyI: ['i', 'i', 'ш', 'ш'],
+  KeyO: ['o', 'o', 'щ', 'щ'],
+  KeyP: ['p', 'p', 'з', 'з'],
+  BracketLeft: ['[', '{', 'х', 'х'],
+  BracketRight: [']', '}', 'ъ', 'ъ'],
+  Backslash: ['\\', '|', 'ё', 'ё'],
+  CapsLock: ['Capslock', 'Capslock', 'Capslock', 'Capslock'],
+  KeyA: ['a', 'a', 'ф', 'ф'],
+  KeyS: ['s', 's', 'ы', 'ы'],
+  KeyD: ['d', 'd', 'в', 'в'],
+  KeyF: ['f', 'f', 'а', 'а'],
+  KeyG: ['g', 'g', 'п', 'п'],
+  KeyH: ['h', 'h', 'р', 'р'],
+  KeyJ: ['j', 'j', 'о', 'о'],
+  KeyK: ['k', 'k', 'л', 'л'],
+  KeyL: ['l', 'l', 'д', 'д'],
+  Semicolon: [';', ':', 'ж', 'ж'],
+  Quote: ["'", '"', 'э', 'э'],
+  Enter: ['Enter', 'Enter', 'Enter', 'Enter'],
+  ShiftLeft: ['Shift', 'Shift', 'Shift', 'Shift'],
+  IntlBackslash: ['`', '~', ']', '['],
+  KeyZ: ['z', 'z', 'я', 'я'],
+  KeyX: ['x', 'x', 'ч', 'ч'],
+  KeyC: ['c', 'c', 'с', 'с'],
+  KeyV: ['v', 'v', 'м', 'м'],
+  KeyB: ['b', 'b', 'и', 'и'],
+  KeyN: ['n', 'n', 'т', 'т'],
+  KeyM: ['m', 'm', 'ь', 'ь'],
+  Comma: [',', '<', 'б', 'б'],
+  Period: ['.', '>', 'ю', 'ю'],
+  Slash: ['/', '?', '/', '?'],
+  ArrowUp: ['&#8593;', '&#8593;', '&#8593;', '&#8593;'],
+  ShiftRight: ['Shift', 'Shift', 'Shift', 'Shift'],
+  ControlLeft: ['Control', 'Control', 'Control', 'Control'],
+  AltLeft: ['Option', 'Option', 'Option', 'Option'],
+  MetaLeft: ['Cmd', 'Cmd', 'Cmd', 'Cmd'],
+  Space: [' ', ' ', ' ', ' '],
+  MetaRight: ['Cmd', 'Cmd', 'Cmd', 'Cmd'],
+  AltRight: ['Option', 'Option', 'Option', 'Option'],
+  ArrowLeft: ['&#8592;', '&#8592;', '&#8592;', '&#8592;'],
+  ArrowDown: ['&#8595;', '&#8595;', '&#8595;', '&#8595;'],
+  ArrowRight: ['&#8594;', '&#8594;', '&#8594;', '&#8594;'],
+
+};
+const elemsUnhandle = [
+  'Tab',
+  'CapsLock',
+  'ShiftLeft',
+  'ControlLeft',
+  'MetaLeft',
+  'MetaRight',
+  'AltLeft',
+  'AltRight',
+  'ShiftRight',
+  'Enter',
+  'Backspace',
+];
+export { keysObj, elemsUnhandle };
