+++ index.js
@@ -1,4 +1,4 @@
 // eslint-disable-next-line import/extensions
-import initLayout from './components/layout.js';
+import initLayout from './components/keyboard.js';
 
 initLayout();
